requirements.txt
google-genai
anthropic
python-telegram-bot[ext]
python-dotenv
chromadb
fastapi
uvicorn[standard]
pydantic
